chore: Use temp_requirements_file fixture
