+++ packages/cli/src/pywrangler/utils.py
@@ -5,8 +5,10 @@ import re
 import shutil
 import subprocess
 import sys
+import tempfile
 import tomllib
-from collections.abc import Mapping
+from collections.abc import Iterator, Mapping
+from contextlib import contextmanager
 from datetime import datetime
 from functools import cache
 from pathlib import Path
@@ -410,3 +412,12 @@ def get_pyodide_index() -> str:
 
 def get_lockfile_path() -> Path:
     return get_project_root() / LOCKFILE_NAME
+
+
+@contextmanager
+def temp_requirements_file(requirements: list[str]) -> Iterator[str]:
+    # Write dependencies to a requirements.txt-style temp file.
+    with tempfile.NamedTemporaryFile(mode="w", suffix=".txt") as temp_file:
+        temp_file.write("\n".join(requirements))
+        temp_file.flush()
+        yield temp_file.name
